Rename example package to ee.jakarta.examples.ai.agent.coursecontentstudio
Align the sample's Java package with the sibling examples
(fraud-detection, docs-agent) under ee.jakarta.examples.ai.agent.
