+++ examples/course-content-studio/src/main/java/ee/jakarta/examples/ai/agent/coursecontentstudio/CourseResource.java
@@ -10,14 +10,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *****************************************************************************/
-package ai.agentic.samples.course;
+package ee.jakarta.examples.ai.agent.coursecontentstudio;
 
-import ai.agentic.samples.course.model.CoursePacket;
-import ai.agentic.samples.course.model.CoursePacketRequest;
-import ai.agentic.samples.course.model.LessonApproved;
-import ai.agentic.samples.course.model.PublishedLesson;
-import ai.agentic.samples.course.model.Quiz;
-import ai.agentic.samples.course.model.QuizQuestion;
+import ee.jakarta.examples.ai.agent.coursecontentstudio.model.CoursePacket;
+import ee.jakarta.examples.ai.agent.coursecontentstudio.model.CoursePacketRequest;
+import ee.jakarta.examples.ai.agent.coursecontentstudio.model.LessonApproved;
+import ee.jakarta.examples.ai.agent.coursecontentstudio.model.PublishedLesson;
+import ee.jakarta.examples.ai.agent.coursecontentstudio.model.Quiz;
+import ee.jakarta.examples.ai.agent.coursecontentstudio.model.QuizQuestion;
 import jakarta.ai.agent.LargeLanguageModel;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.enterprise.event.Event;
