+++ examples/course-content-studio/src/main/java/ee/jakarta/examples/ai/agent/coursecontentstudio/RestApplication.java
@@ -10,16 +10,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *****************************************************************************/
-package ai.agentic.samples.course.model;
+package ee.jakarta.examples.ai.agent.coursecontentstudio;
 
-import java.util.List;
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
 
 /**
- * A quiz: the typed result of an LLM query. Using a record (instead of parsing
- * a raw String) lets the agent ask for {@code Quiz.class} and receive structured
- * data straight from Jakarta JSON Binding.
- *
- * @param questions the quiz questions
+ * Activates JAX-RS under {@code /api}.
  */
-public record Quiz(List<QuizQuestion> questions) {
+@ApplicationPath("/api")
+public class RestApplication extends Application {
 }
